🚚 move: pages into components
