+++ src/pages/_offline/index.page.tsx
@@ -1,30 +1,12 @@
-import type { GetStaticProps, InferGetStaticPropsType, NextPage } from "next";
-import { FaRegTired } from "react-icons/fa";
+import type { GetStaticProps, NextPage } from "next";
 
 import { fetchCategories, fetchConfig } from "@/api";
-import { HtmlHeadNoIndex } from "@/components/functions/meta";
-import { RootLayout } from "@/components/layouts/RootLayout";
-import type { TCategory, TConfig } from "@/types";
+import type { OfflineProps } from "@/components/pages/Offline";
+import { Offline } from "@/components/pages/Offline";
 
-type OfflineProps = InferGetStaticPropsType<typeof getStaticProps>;
+const OfflinePage: NextPage<OfflineProps> = (props) => <Offline {...props} />;
 
-const Offline: NextPage<OfflineProps> = ({ config, categories }) => (
-  <RootLayout config={config} categories={categories}>
-    <HtmlHeadNoIndex />
-    <div className="mt-8 flex flex-col items-center justify-center gap-12" role="alert">
-      <FaRegTired className="h-32 w-32 text-gray-500" />
-      <div className="mx-auto flex flex-col gap-8 text-center">
-        <h1>オフラインページ</h1>
-        <p>インターネットに接続のうえ、ご利用ください。</p>
-      </div>
-    </div>
-  </RootLayout>
-);
-
-type StaticProps = {
-  config: TConfig;
-  categories: TCategory[];
-};
+type StaticProps = OfflineProps;
 
 export const getStaticProps: GetStaticProps<StaticProps> = async () => {
   const config = await fetchConfig();
@@ -37,4 +19,4 @@ export const getStaticProps: GetStaticProps<StaticProps> = async () => {
     },
   };
 };
-export default Offline;
+export default OfflinePage;
