--- src/pages/articles/categories/index.test.tsx
@@ -1,41 +1,0 @@
-import { mockCategories, mockConfig, mockPickup, mockPopularArticles } from "@mocks/data";
-import { render, screen } from "jest/test-utils";
-import { server } from "mocks/msw/server";
-
-import { formatPageTitle } from "@/utils/formatter";
-
-import Categories from "./index.page";
-
-beforeAll(() => server.listen());
-afterAll(() => server.close());
-
-jest.mock("next/head", () => {
-  return {
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    __esModule: true,
-    default: ({ children }: { children: Array<React.ReactElement> }) => {
-      return <>{children}</>;
-    },
-  };
-});
-
-describe("pages/articles/categories", () => {
-  const mockCategoryList = Object.values(mockCategories);
-
-  it("OK: 初期レンダリング", async () => {
-    const { unmount } = render(
-      <Categories
-        categories={mockCategoryList}
-        config={mockConfig}
-        pickup={mockPickup}
-        popularArticles={mockPopularArticles}
-      />
-    );
-    const h1 = screen.getByRole("heading", { level: 1 });
-    const expectedHeading = `カテゴリー一覧`;
-    expect(h1).toHaveTextContent(expectedHeading);
-    const expectedTitle = formatPageTitle(expectedHeading, mockConfig.siteTitle);
-    expect(document.title).toBe(expectedTitle);
-    unmount();
-  });
-});
