+++ src/pages/articles/tags/[slug].page.tsx
@@ -2,43 +2,14 @@ import type { GetStaticPaths, GetStaticProps, InferGetStaticPropsType, NextPage 
 import type { ParsedUrlQuery } from "node:querystring";
 
 import { fetchArticles, fetchCategories, fetchConfig, fetchTag, fetchTags } from "@/api";
-import { HtmlHeadBase } from "@/components/functions/meta";
-import DefaultLayout from "@/components/layouts/DefaultLayout";
-import { ArticleSWRContainer } from "@/components/organisms/ArticleSWRContainer";
+import type { TagsProps } from "@/components/pages/articles/tags";
+import { Tags } from "@/components/pages/articles/tags";
 import { getPickupArticles, getPopularArticles } from "@/services/article";
-import type { TArticleListResponse, TCategory, TConfig, TPickup, TRankedArticle, TTag } from "@/types";
-import { formatPageTitle, formatPageUrl } from "@/utils/formatter";
-import { getBackLinks, urlTable } from "@/utils/paths/url";
 
 type Props = InferGetStaticPropsType<typeof getStaticProps>;
 
-const Tags: NextPage<Props> = ({ data, tag, config, categories, pickup, popularArticles }) => {
-  const { siteTitle, host } = config;
-  const heading = `タグ：${tag.name}`;
-  const pageTitle = formatPageTitle(heading, siteTitle);
-  const url = formatPageUrl(`${urlTable.tags}/${tag.slug}`, host);
-  const backLinks = getBackLinks([urlTable.home, urlTable.categories]);
-  const queryOptions = { filters: `tags[contains]${tag.id}` };
-
-  return (
-    <DefaultLayout
-      config={config}
-      pageTitle={pageTitle}
-      url={url}
-      backLinks={backLinks}
-      categories={categories}
-      pickup={pickup}
-      popularArticles={popularArticles}
-    >
-      <HtmlHeadBase indexUrl={host} pageTitle={pageTitle} url={url} />
-      <div className="mb-8">
-        <h1>{heading}</h1>
-      </div>
-      <div className="min-h-screen w-full">
-        <ArticleSWRContainer fallbackData={data} queryOptions={queryOptions} />
-      </div>
-    </DefaultLayout>
-  );
+const TagsPage: NextPage<Props> = (props) => {
+  return <Tags {...props} />;
 };
 
 interface Params extends ParsedUrlQuery {
@@ -54,14 +25,7 @@ export const getStaticPaths: GetStaticPaths<Params> = async () => {
   return { paths, fallback: "blocking" };
 };
 
-type StaticProps = {
-  tag: TTag;
-  categories: TCategory[];
-  config: TConfig;
-  data: TArticleListResponse;
-  pickup: TPickup;
-  popularArticles: TRankedArticle[];
-};
+type StaticProps = TagsProps;
 
 export const getStaticProps: GetStaticProps<StaticProps, Params> = async ({ params }) => {
   const slug = params?.slug;
@@ -90,4 +54,4 @@ export const getStaticProps: GetStaticProps<StaticProps, Params> = async ({ para
   };
 };
 
-export default Tags;
+export default TagsPage;
