+++ src/pages/index.page.tsx
@@ -1,43 +1,18 @@
-import type { GetStaticProps, InferGetStaticPropsType } from "next";
+import type { GetStaticProps, NextPage } from "next";
 
 import { fetchArticles, fetchCategories, fetchConfig } from "@/api";
-import { HtmlHeadBase } from "@/components/functions/meta";
-import HomeLayout from "@/components/layouts/HomeLayout";
-import { ArticleSWRContainer } from "@/components/organisms/ArticleSWRContainer";
+import type { HomeProps } from "@/components/pages/Home";
+import { Home } from "@/components/pages/Home";
 import { getPickupArticles, getPopularArticles } from "@/services/article";
-import type { TArticleListResponse, TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
 import { generatedRssFeed } from "@/utils/rss/rss";
 
-type Props = InferGetStaticPropsType<typeof getStaticProps>;
-
-const Home = ({ data, config, pickup, categories, popularArticles }: Props) => {
-  const { siteTitle: title, host } = config;
-
-  return (
-    <HomeLayout
-      pickup={pickup}
-      url={host}
-      pageTitle={title}
-      config={config}
-      categories={categories}
-      popularArticles={popularArticles}
-    >
-      <HtmlHeadBase indexUrl={host} siteTitle={title} />
-      <div className="mb-8">
-        <h1>レシピ一覧</h1>
-      </div>
-      <ArticleSWRContainer fallbackData={data} />
-    </HomeLayout>
-  );
+const HomePage: NextPage<HomeProps> = (props) => {
+  return <Home {...props} />;
 };
 
-type StaticProps = {
-  data: TArticleListResponse;
-  categories: TCategory[];
-  config: TConfig;
-  pickup: TPickup;
-  popularArticles: TRankedArticle[];
-};
+export default HomePage;
+
+type StaticProps = HomeProps;
 
 export const getStaticProps: GetStaticProps<StaticProps> = async () => {
   const [config, categories, data, pickup, popularArticles] = await Promise.all([
@@ -60,5 +35,3 @@ export const getStaticProps: GetStaticProps<StaticProps> = async () => {
     },
   };
 };
-
-export default Home;
