+++ src/pages/search/index.page.tsx
@@ -1,58 +1,15 @@
-import type { GetStaticProps, InferGetStaticPropsType, NextPage } from "next";
-import { useRouter } from "next/router";
+import type { GetStaticProps, NextPage } from "next";
 
 import { fetchCategories, fetchConfig } from "@/api";
-import { HtmlHeadBase } from "@/components/functions/meta";
-import DefaultLayout from "@/components/layouts/DefaultLayout";
-import { ArticleSWRContainer } from "@/components/organisms/ArticleSWRContainer";
+import type { SearchProps } from "@/components/pages/Search";
+import { Search } from "@/components/pages/Search";
 import { getPickupArticles, getPopularArticles } from "@/services/article";
-import type { TCategory, TConfig, TPickup, TQueryOptions, TRankedArticle } from "@/types";
-import { formatPageTitle, formatPageUrl } from "@/utils/formatter";
-import { getBackLinks, urlTable } from "@/utils/paths/url";
 
-type Props = InferGetStaticPropsType<typeof getStaticProps>;
-
-const Search: NextPage<Props> = ({ config, categories, pickup, popularArticles }) => {
-  const router = useRouter();
-  const q = router.query.q?.toString();
-  const { siteTitle, host } = config;
-  const heading = `検索結果：${q ?? ""}`;
-  const pageTitle = formatPageTitle(heading, siteTitle);
-  const url = formatPageUrl(`${urlTable.search}/q=${q ?? ""}`, host);
-  const backLinks = getBackLinks([urlTable.home, urlTable.categories]);
-  const queryOptions: TQueryOptions = { q };
-
-  return (
-    <DefaultLayout
-      config={config}
-      pageTitle={pageTitle}
-      url={url}
-      backLinks={backLinks}
-      categories={categories}
-      pickup={pickup}
-      popularArticles={popularArticles}
-    >
-      <HtmlHeadBase indexUrl={host} pageTitle={pageTitle} url={url} />
-      <div className="mb-8">
-        <h1>{heading}</h1>
-      </div>
-      <div className="w-full">
-        {q ? (
-          <ArticleSWRContainer queryOptions={queryOptions} />
-        ) : (
-          <div className="mt-16 flex justify-center">検索語句を入力してください。</div>
-        )}
-      </div>
-    </DefaultLayout>
-  );
+const SearchPage: NextPage<SearchProps> = (props) => {
+  return <Search {...props} />;
 };
 
-type StaticProps = {
-  config: TConfig;
-  categories: TCategory[];
-  pickup: TPickup;
-  popularArticles: TRankedArticle[];
-};
+type StaticProps = SearchProps;
 
 export const getStaticProps: GetStaticProps<StaticProps> = async () => {
   const [config, categories, pickup, popularArticles] = await Promise.all([
@@ -72,4 +29,4 @@ export const getStaticProps: GetStaticProps<StaticProps> = async () => {
   };
 };
 
-export default Search;
+export default SearchPage;
